+++ src/dune_engine/source_tree.ml
@@ -476,6 +476,23 @@ end = struct
       (visited, init)
   end
 
+  let ensure_dune_project_file_exists =
+    Memo.create "ensure-dune-project-file-exists"
+      ~input:(module Dune_project)
+      (fun project ->
+        let open Memo.Build.O in
+        let+ exists =
+          Dune_project.file project |> Path.source |> Fs_memo.file_exists
+        in
+        if not exists then
+          User_warning.emit
+            ~is_error:!Clflags.require_dune_project_file
+            [ Pp.text
+                "No dune-project file has been found. A default one is assumed \
+                 but the project might break when dune is upgraded. Please \
+                 create a dune-project file."
+            ])
+
   let dune_file ~(dir_status : Sub_dirs.Status.t) ~path ~files ~project =
     let file_exists =
       if dir_status = Data_only then
@@ -511,6 +528,8 @@ end = struct
       | None, Some (path, _) -> Some path
     in
     Memo.Build.Option.map file ~f:(fun file ->
+        let open Memo.Build.O in
+        let* () = Memo.exec ensure_dune_project_file_exists project in
         let file_exists = Option.is_some file_exists in
         let from_parent = Option.map from_parent ~f:snd in
         Dune_file.load file ~file_exists ~project ~from_parent)