+++ test/blackbox-tests/test-cases/github4684.t
@@ -3,6 +3,9 @@ build failure:
 
   $ echo "module X = Root.Unix" > main.ml
   $ touch main.opam
+  $ cat >dune-project <<EOF
+  > (lang dune 2.8)
+  > EOF
   $ cat >dune <<EOF
   > (library
   >  (name main)